Renamed and Culled Logic to core required.
