--- src/lib/index.js
@@ -1,2 +1,0 @@
-// Reexport your entry components here
-export {buildURL, validateSearchParams, validatedSearchParamsStore} from './skSearchParams.js'
