+++ src/lib/index.ts
@@ -1,0 +1,3 @@
+// Reexport your entry components here
+export { skRoutes as createURLGenerator } from './skRoutes.js';
+export type { RouteConfig } from './skRoutes.js';
